Added `get_css_position` filter
Often the image has to be used as a background-image for the container
which size is unknown. Usual way of doing it is setting background image
to the container and using `background-size: cover;`. However if you do
that - focal point position won't be respected. With this template tag
you can now do the following:

```
{% load filer_iamge_tags %}
<div class="hero"
    style="background-image: url({{ image.url }}; background-position: {{ image|get_css_position }};">
    I'm a hero unit
</div>
```

And the image would be positioned correctly.